CaseRef changed to long

+++ src/main/java/uk/gov/ons/census/action/builders/CaseSelectedBuilder.java
@@ -36,7 +36,7 @@ public class CaseSelectedBuilder {
     responseManagementEvent.getPayload().setFieldCaseSelected(fieldCaseSelected);
 
     fieldCaseSelected.setActionRuleId(actionRuleId.toString());
-    fieldCaseSelected.setCaseRef(Integer.parseInt(caseRef));
+    fieldCaseSelected.setCaseRef(Long.parseLong(caseRef));
 
     return responseManagementEvent;
   }
