+++ src/main/java/uk/gov/ons/census/action/model/repository/CaseRepository.java
@@ -4,5 +4,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import uk.gov.ons.census.action.model.entity.Case;
 
-public interface CaseRepository
-    extends JpaRepository<Case, Integer>, JpaSpecificationExecutor<Case> {}
+public interface CaseRepository extends JpaRepository<Case, Long>, JpaSpecificationExecutor<Case> {}
